# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the University of Jordan Volunteer Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  volunteer: 'volunteer',
} as const;

export interface User {
  id: number;
  phone: string;
  role: UserRole;
  fullName: string;
  /** @nullable */
  major?: string | null;
  /** @nullable */
  university?: string | null;
  acceptedRequestsCount?: number;
  completedRequestsCount?: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  student: 'student',
  volunteer: 'volunteer',
} as const;

export interface RegisterInput {
  phone: string;
  password: string;
  role: RegisterInputRole;
  fullName: string;
  major?: string;
  university: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export type HelpRequestStatus = typeof HelpRequestStatus[keyof typeof HelpRequestStatus];


export const HelpRequestStatus = {
  open: 'open',
  accepted: 'accepted',
  completed: 'completed',
} as const;

export interface HelpRequest {
  id: number;
  studentId: number;
  studentName: string;
  /** @nullable */
  studentMajor?: string | null;
  studentPhone: string;
  /** @nullable */
  studentUniversity?: string | null;
  /** @nullable */
  volunteerId?: number | null;
  /** @nullable */
  volunteerName?: string | null;
  subjectName: string;
  isExam: boolean;
  /** @nullable */
  examDate?: string | null;
  /** @nullable */
  examLocation?: string | null;
  status: HelpRequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface HelpRequestInput {
  subjectName: string;
  isExam: boolean;
  examDate?: string;
  examLocation?: string;
  notes?: string;
}

export interface RequestsSummary {
  totalOpen: number;
  totalAccepted: number;
  totalCompleted: number;
  recentRequests: HelpRequest[];
}

export type ListRequestsParams = {
status?: ListRequestsStatus;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  open: 'open',
  accepted: 'accepted',
  completed: 'completed',
} as const;