# lib/api-zod/src/generated/types/helpRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the University of Jordan Volunteer Platform
 * OpenAPI spec version: 0.1.0
 */
import type { HelpRequestStatus } from './helpRequestStatus';

export interface HelpRequest {
  id: number;
  studentId: number;
  studentName: string;
  /** @nullable */
  studentMajor?: string | null;
  studentPhone: string;
  /** @nullable */
  studentUniversity?: string | null;
  /** @nullable */
  volunteerId?: number | null;
  /** @nullable */
  volunteerName?: string | null;
  subjectName: string;
  isExam: boolean;
  /** @nullable */
  examDate?: Date | null;
  /** @nullable */
  examLocation?: string | null;
  status: HelpRequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}