# lib/api-zod/src/generated/types/requestsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the University of Jordan Volunteer Platform
 * OpenAPI spec version: 0.1.0
 */
import type { HelpRequest } from './helpRequest';

export interface RequestsSummary {
  totalOpen: number;
  totalAccepted: number;
  totalCompleted: number;
  recentRequests: HelpRequest[];
}