# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the University of Jordan Volunteer Platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  phone: string;
  role: UserRole;
  fullName: string;
  /** @nullable */
  major?: string | null;
  /** @nullable */
  university?: string | null;
  acceptedRequestsCount?: number;
  completedRequestsCount?: number;
  createdAt: Date;
}